requirements.txt
langchain==0.1.0
langchain-experimental==0.0.49
openai==1.7.2
pandas==2.1.4
python-dotenv==1.0.0
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
snowflake-sqlalchemy==1.5.1
